python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ortools==9.8.3296
geopy==2.4.0
pydantic==2.5.0
python-multipart==0.0.6
folium==0.15.1
matplotlib==3.8.2
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0 